Move mechanic logic to Skill module
